feat(decorator): Stage 3 decorator 인프라 + 실패 테스트 추가
- RuntimeHelpers packed struct(u16) → packed struct(u32) 확장
- es_decorator 플래그 추가 (TC39 Stage 3 decorator 런타임 필요 시 설정)
- tslib 기반 __esDecorate/__runInitializers/__setFunctionName/__propKey 런타임 헬퍼 추가
- emitter에 es_decorator 헬퍼 주입 조건 추가
- e2eStage3Decorator() 테스트 헬퍼 추가
- 11개 실패하는 Stage 3 테스트 작성 (class/method/field/getter/setter/accessor)

+++ src/transpile.zig
@@ -306,7 +306,7 @@ pub fn transpileWithCallback(
 
     // 7. 런타임 헬퍼 prepend
     const rh = transformer.runtime_helpers;
-    const has_helpers = @as(u16, @bitCast(rh)) != 0;
+    const has_helpers = @as(u32, @bitCast(rh)) != 0;
     const output = if (has_helpers) blk: {
         var buf: std.ArrayList(u8) = .empty;
         rt.appendRuntimeHelpers(&buf, arena_alloc, rh, options.minify_whitespace, transformer.runtime_es5_compat) catch
